+++ src/status.rs
@@ -3,14 +3,14 @@ use crate::sdl2::pixels::Color;
 use crate::system::System;
 use crate::ONE_SECOND;
 
-const STATUS_UPDATE_TIME: u64 = ONE_SECOND / 60;
+const STATUS_UPDATE_TIME: u32 = ONE_SECOND / 60;
 
 pub struct Status {
     text: String,
     // When it reaches y_limit, the status should be removed.
     y_pos: i32,
     y_limit: i32,
-    duration: u64,
+    duration: u32,
     color: Color,
 }
 
@@ -25,7 +25,7 @@ impl<'a> Status {
         }
     }
 
-    pub fn update(&mut self, elapsed: u64) {
+    pub fn update(&mut self, elapsed: u32) {
         self.duration += elapsed;
         while self.duration > STATUS_UPDATE_TIME && self.y_pos < self.y_limit {
             self.duration -= STATUS_UPDATE_TIME;
